# archived/cs-4000/v1/templates/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CS-4000 - Slim PHP",
    "description": "Move your app forward with the CS-4000 Slim PHP API",
    "version": "1.0.0"
  },
  "host": "api.trendlycreations.com",
  "schemes": [
    "http"
  ],
  "basePath": "/v1",
  "produces": [
    "application/json"
  ],
  "paths": {
    "/clients": {
      "get": {
        "summary": "All Clients",
        "description": "The Clients endpoint returns information about all clients in the DB.",
        "tags": [
          "Clients"
        ],
        "responses": {
          "200": {
            "description": "An array of clients",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Client"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/clients/{id}": {
      "get": {
        "summary": "Show Client",
        "description": "The Clients endpoint returns information about a client with the specified id in the DB.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "First name of client.",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Clients"
        ],
        "responses": {
          "200": {
            "description": "A client",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Client"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "summary": "Update Client",
        "description": "The Update Client endpoint updates the clients data and then returns the changed client",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of client.",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "first_name",
            "in": "query",
            "description": "First name of client.",
            "required": true,
            "type": "string"
          },
          {
            "name": "last_name",
            "in": "query",
            "description": "Last name of client.",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "query",
            "type": "string",
            "format": "uuid",
            "description": "Email of client.",
            "required": true
          },
          {
            "name": "company",
            "in": "query",
            "type": "string",
            "description": "Company of client.",
            "required": true
          },
          {
            "name": "project",
            "in": "query",
            "type": "string",
            "description": "Project of client.",
            "required": true
          },
          {
            "name": "phone",
            "in": "query",
            "type": "string",
            "description": "Phone number of client.",
            "required": true
          }
        ],
        "tags": [
          "Clients"
        ],
        "responses": {
          "200": {
            "description": "Updated Clients data",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Client",
        "description": "The Delete Client endpoint returns data about the deleted client",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of client",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Clients"
        ],
        "responses": {
          "200": {
            "description": "Deleted clients data",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/clients/": {
      "post": {
        "summary": "Create Client",
        "description": "Creates a new Client and then return the newly created client.",
        "parameters": [
          {
            "name": "first_name",
            "in": "query",
            "description": "First name of client.",
            "required": true,
            "type": "string"
          },
          {
            "name": "last_name",
            "in": "query",
            "description": "Last name of client.",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "query",
            "type": "string",
            "format": "uuid",
            "description": "Email of client.",
            "required": true
          },
          {
            "name": "company",
            "in": "query",
            "type": "string",
            "description": "Company of client.",
            "required": true
          },
          {
            "name": "project",
            "in": "query",
            "type": "string",
            "description": "Project of client.",
            "required": true
          },
          {
            "name": "phone",
            "in": "query",
            "type": "string",
            "description": "Phone number of client.",
            "required": true
          }
        ],
        "tags": [
          "Clients"
        ],
        "responses": {
          "201": {
            "description": "A Client",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Client"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Client": {
      "properties": {
        "product_id": {
          "type": "string",
          "description": "Unique identifier representing a specific product for a given latitude & longitude. For example, uberX in San Francisco will have a different product_id than uberX in Los Angeles."
        },
        "description": {
          "type": "string",
          "description": "Description of product."
        },
        "display_name": {
          "type": "string",
          "description": "Display name of product."
        },
        "capacity": {
          "type": "string",
          "description": "Capacity of product. For example, 4 people."
        },
        "image": {
          "type": "string",
          "description": "Image URL representing the product."
        }
      }
    },
    "Error": {
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        }
      }
    }
  }
}